// frontend/src/generated/types/emailValidRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface EmailValidRequest {
  sysId?: string;
  linkSysId?: string;
  encodeSysId?: string;
  validCode?: string;
  email?: string;
  userId?: string;
}